rename landing in application and application in application_agent
