+++ app/controllers/ehpads_controller.rb
@@ -1,5 +1,4 @@
 class EhpadsController < ApplicationController
-  layout 'landing'
   before_action :set_variables, only: [:search]
 
   def set_variables
