+++ app/controllers/static_pages_controller.rb
@@ -1,5 +1,4 @@
 class StaticPagesController < ApplicationController
-  layout 'landing'
 
   def disclaimer; end
