+++ app/controllers/user_auth_controller.rb
@@ -1,7 +1,6 @@
 class UserAuthController < ApplicationController
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
-  layout 'landing'
   before_action :authenticate_user!
   before_action :set_paper_trail_whodunnit
