+++ app/views/common/_head_agent.html.slim
@@ -1,0 +1,4 @@
+= render 'common/meta'
+= render 'common/env'
+= stylesheet_pack_tag 'application_agent', media: 'all', 'data-turbolinks-track': 'reload'
+= javascript_pack_tag 'application_agent', 'data-turbolinks-track': 'reload'
