--- app/views/common/_head_landing.html.slim
@@ -1,4 +1,0 @@
-= render 'common/meta'
-= render 'common/env'
-= stylesheet_pack_tag 'landing', media: 'all', 'data-turbolinks-track': 'reload'
-= javascript_pack_tag 'landing', 'data-turbolinks-track': 'reload'
