--- app/views/layouts/landing.html.slim
@@ -1,27 +1,0 @@
-doctype html
-html lang="fr"
-  head
-    = render 'common/head_landing'
-  body class="#{agents_or_users_body_class}"
-    div class="#{ params[:controller].include?('users/') ? '' : 'bg-primary'}"
-      - if current_user.present?
-        = render "common/header_user_logged"
-      - else
-        = render 'common/header'
-
-      .container
-        = render 'layouts/flash'
-        - if content_for :title
-          .row
-            .col-md-12.d-flex.justify-content-between.align-items-center
-              .my-4= yield :title
-              - if content_for :breadcrumb
-                div
-                  = yield :breadcrumb
-
-      = yield :header
-
-    div class="#{ params[:controller].include?('users/') ? 'container' : 'mt-3'}"
-      = yield
-    #modal-holder
-    = render 'common/footer'
