+++ app/webpacker/stylesheets/application.scss
@@ -9,35 +9,284 @@
 @import "./mixins/buttons";
 @import "./mixins/background-variant";
 
-// Plugins
-@import "actiontext";
-@import '@fullcalendar/core/main';
-@import '@fullcalendar/daygrid/main';
-@import '@fullcalendar/timegrid/main';
-@import '@fullcalendar/list/main';
-@import "jquery-datetimepicker/build/jquery.datetimepicker.min";
+@import "./components/backgrounds";
+@import "./components/types";
+@import "./components/search_form";
+@import "./components/forms";
+@import "./components/utilities";
+@import "./components/steps";
 
 // Structure
 @import "./structure/general";
-@import "./structure/left-menu";
 @import "./structure/topbar";
 @import "./structure/footer";
 @import "./structure/page-head";
 @import "./structure/authentication";
-@import "./structure/right_sidebar";
 
-// Components
-@import "./components/backgrounds";
-@import "./components/calendar";
-@import "./components/forms";
-@import "./components/breadcrumb";
-@import "./components/dropdown";
-@import "./components/motifs";
-@import "./components/nav";
-@import "./components/tables";
-@import "./components/types";
-@import "./components/cards";
-@import "./components/timeline";
-@import "./components/progress_bars";
-@import "./components/recurrence";
-@import "./components/utilities";
+h1 {
+  font-size: 3rem;
+  font-weight: bold;
+  color: white;
+}
+
+h2{
+  color: $primary;
+  font-weight: bold;
+  font-size: 2rem;
+  padding-bottom: 1rem;
+}
+
+a {
+  transition: all 0.3s;
+  display: inline-block;
+}
+
+p {
+  line-height: 1.7rem;
+}
+
+section header {
+  margin-bottom: 70px;
+}
+
+
+/* General Media Query ---------------------------------- */
+@media (max-width: 991px) {
+  h2 {
+    text-align: center;
+  }
+
+  .container {
+    max-width: 90%;
+  }
+
+  section {
+    padding: 100px 0;
+  }
+
+  .img-fluid {
+    max-width: 60%;
+  }
+
+  .content {
+    .col-lg-4{
+      order: 1;
+    }
+    .col-lg-8{
+      order: 2;
+    }
+  }
+}
+
+/*
+* ==========================================================
+*     NAVBAR
+* ==========================================================
+*/
+
+nav.navbar {
+  background: none;
+  transition: all 0.5s;
+  .navbar-brand {
+    font-weight: 800;
+    font-size: 1.7rem;
+    color: #222;
+  }
+  a.nav-link {
+    color: #333;
+    font-size: 1.1rem;
+    &:hover, &.active {
+      color: $primary;
+    }
+  }
+  .navbar-btn {
+    margin-left: 15px;
+  }
+  .navbar-toggler {
+    outline: none;
+    span {
+      width: 25px;
+      margin: 5px;
+      display: block;
+      height: 2px;
+      background: #333;
+      -webkit-transition: all 0.2s;
+      transition: all 0.2s;
+    }
+    &.active span {
+      &:first-of-type {
+        -webkit-transform: rotate(45deg) translate(10px);
+        transform: rotate(45deg) translate(10px);
+      }
+      &:last-of-type {
+        -webkit-transform: rotate(-45deg) translate(10px);
+        transform: rotate(-45deg) translate(10px);
+      }
+      &:nth-of-type(2) {
+        opacity: 0;
+      }
+    }
+  }
+}
+
+/* Navbar Mediaquery ----------------------------------- */
+@media (max-width: 1199px) {
+  nav.navbar .navbar-btn {
+    margin-left: 0;
+  }
+}
+
+@media (min-width: 992px) {
+  nav.navbar {
+    padding-left: 3rem;
+    padding-right: 3rem;
+  }
+}
+
+/*
+* ==========================================================
+*     HERO SECTION
+* ==========================================================
+*/
+
+section {
+  &.hero {
+    p {
+      margin: 20px 0;
+    }
+  }
+  &.our-mission p {
+    margin: 30px 0;
+  }
+
+  &.who i {
+    font-size: 4em;
+  }
+}
+
+
+/*
+* ==========================================================
+*     FOOTER
+* ==========================================================
+*/
+
+footer.main-footer {
+  padding: 3rem 0 0;
+  overflow-x: hidden;
+  a {
+    color: inherit;
+    text-decoration: none;
+  }
+  .brand {
+    font-weight: 800;
+    font-size: 1.7rem;
+    color: #222;
+    margin-bottom: 20px;
+  }
+  ul.contact-info {
+    color: #444;
+    a {
+      margin-bottom: 5px;
+      &:hover {
+        color: $primary;
+      }
+    }
+  }
+  h5 {
+    margin-bottom: 30px;
+  }
+  ul {
+    &.links {
+      color: #888;
+      a {
+        margin-bottom: 8px;
+        &:hover {
+          color: $primary;
+        }
+      }
+    }
+    &.social-icons {
+      margin-top: 15px;
+      margin-bottom: 50px;
+      a {
+        width: 26px;
+        height: 26px;
+        line-height: 26px;
+        background: #222;
+        color: #fff;
+        text-align: center;
+        border-radius: 50%;
+        font-size: 0.8rem;
+        &:hover {
+          background: $primary;
+        }
+      }
+    }
+  }
+  .copyrights {
+    margin-top: 150px;
+    .container {
+      padding: 30px 0;
+      border-top: 1px solid #ddd;
+    }
+    p {
+      margin-bottom: 0;
+    }
+    a:hover {
+      color: $primary;
+    }
+  }
+  .template {
+    font-size: 0.8rem;
+    color: rgb(136, 136, 136);
+  }
+}
+
+/* Footer Mediaquery -------------------------- */
+@media (max-width: 991px) {
+  footer.main-footer {
+    padding: 100px 0 0;
+    .copyrights {
+      margin-top: 100px;
+      div[class*="col-"] {
+        text-align: center !important;
+      }
+    }
+  }
+}
+
+
+/* =========================================
+   THEMING OF BOOTSTRAP COMPONENTS
+   ========================================= */
+/*
+ * 1. NAVBAR
+ */
+
+.navbar {
+  padding: 0.5rem 1rem;
+}
+
+.navbar-brand {
+  display: inline-block;
+  padding-top: 0.3125rem;
+  padding-bottom: 0.3125rem;
+  margin-right: 1rem;
+  font-size: 1.25rem;
+}
+
+.navbar-toggler {
+  padding: 0.25rem 0.75rem;
+  font-size: 1.25rem;
+  line-height: 1;
+  border: 1px solid transparent;
+  border-radius: 50px;
+}
+
+.img-circle {
+  margin-top: -70px;
+  @include media-breakpoint-down(md) {
+    margin-top: 0;
+  }
+}
