+++ app/webpacker/stylesheets/application_agent.scss
@@ -1,0 +1,42 @@
+@import "~bootstrap/scss/functions";
+@import "~bootstrap/scss/variables";
+@import "./_variables";
+@import "./fontawesome_imports";
+
+@import "~bootstrap/scss/bootstrap";
+@import "~select2/dist/css/select2";
+@import "~select2-bootstrap4-theme/dist/select2-bootstrap4.min";
+@import "./mixins/buttons";
+@import "./mixins/background-variant";
+
+// Plugins
+@import "actiontext";
+@import '@fullcalendar/core/main';
+@import '@fullcalendar/daygrid/main';
+@import '@fullcalendar/timegrid/main';
+@import '@fullcalendar/list/main';
+@import "jquery-datetimepicker/build/jquery.datetimepicker.min";
+
+// Structure
+@import "./structure/general";
+@import "./structure/left-menu";
+@import "./structure/topbar";
+@import "./structure/footer";
+@import "./structure/page-head";
+@import "./structure/right_sidebar";
+
+// Components
+@import "./components/backgrounds";
+@import "./components/calendar";
+@import "./components/forms";
+@import "./components/breadcrumb";
+@import "./components/dropdown";
+@import "./components/motifs";
+@import "./components/nav";
+@import "./components/tables";
+@import "./components/types";
+@import "./components/cards";
+@import "./components/timeline";
+@import "./components/progress_bars";
+@import "./components/recurrence";
+@import "./components/utilities";
